requirements.txt
# pip install -r requirements.txt
langchain
langchain_community
langgraph
langchain-openai
tavily-python
python-dotenv